+++ lib/danger/commands/init.rb
@@ -272,7 +272,7 @@ module Danger
       ui.header "Useful info"
       ui.say "- One of the best ways to test out new rules locally is via " + "bundle exec danger pr".yellow + "."
       ui.pause 0.6
-      ui.say "- You can have Danger output all of its variables to the console via the " + "--verbose".yellow + " option."
+      ui.say "- You can have Danger output all of her variables to the console via the " + "--verbose".yellow + " option."
       ui.pause 0.6
       ui.say "- You can look at the following Dangerfiles to get some more ideas:"
       ui.pause 0.6
@@ -286,8 +286,8 @@ module Danger
       ui.pause 0.6
 
       ui.say "And you're good to go. Danger is a collaboration between Orta Therox, Gem 'Danger' McShane and Felix Krause."
-      ui.say "If you like it, let others know. If you want to know more, follow " + "@orta".yellow + " and " + "@KrauseFx".yellow + " on Twitter."
-      ui.say "If you don't like it, help us improve it! xxx"
+      ui.say "If you like Danger, let others know. If you want to know more, follow " + "@orta".yellow + " and " + "@KrauseFx".yellow + " on Twitter."
+      ui.say "If you don't like Danger, help us improve the project! xxx"
     end
 
     def darwin?