+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## master
 
 * Add your own contributions to the next release on the line below this, please include your name too. Please don't set a new version if you are the first to make the section for `master`.
+
+## 5.4.0
+
 * Add support for VSTS - [@petester42](https://github.com/petester42)
 
 ## 5.3.5